RDCC-958 added missing tests for any untested classes, moved classes into correct packages
