+++ config/owasp/suppressions.xml
@@ -312,6 +312,13 @@
     </suppress>
     <suppress>
         <notes><![CDATA[
+   file name: checkstyle-8.10.1.jar
+   ]]></notes>
+        <packageUrl regex="true">^pkg:maven/com\.puppycrawl\.tools/checkstyle@.*$</packageUrl>
+        <vulnerabilityName>CVE-2019-10782</vulnerabilityName>
+    </suppress>
+    <suppress>
+        <notes><![CDATA[
    file name: tomcat-embed-core-9.0.17.jar
    ]]></notes>
         <packageUrl regex="true">^pkg:maven/org\.apache\.tomcat\.embed/tomcat-embed-core@.*$</packageUrl>
