--- src/main/java/uk/gov/hmcts/reform/professionalapi/controller/request/Jurisdiction.java
@@ -1,13 +1,0 @@
-package uk.gov.hmcts.reform.professionalapi.controller.request;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@Setter
-@NoArgsConstructor
-public class Jurisdiction {
-
-    private String id;
-}
