+++ src/main/java/uk/gov/hmcts/reform/professionalapi/controller/request/JurisdictionUserCreationRequest.java
@@ -5,6 +5,7 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import java.util.List;
 import lombok.Getter;
 import lombok.Setter;
+import uk.gov.hmcts.reform.professionalapi.domain.Jurisdiction;
 
 @Setter
 @Getter
