+++ src/main/java/uk/gov/hmcts/reform/professionalapi/controller/request/NewUserCreationRequest.java
@@ -8,6 +8,7 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
 import org.apache.commons.lang.StringUtils;
+import uk.gov.hmcts.reform.professionalapi.domain.Jurisdiction;
 
 @Getter
 @Setter
