--- src/main/java/uk/gov/hmcts/reform/professionalapi/controller/response/UserProfile.java
@@ -1,15 +1,0 @@
-package uk.gov.hmcts.reform.professionalapi.controller.response;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class UserProfile {
-
-    private String idamId;
-    private String email;
-    private String firstName;
-    private String lastName;
-    private IdamStatus idamStatus;
-}
