+++ src/main/java/uk/gov/hmcts/reform/professionalapi/domain/Jurisdiction.java
@@ -1,0 +1,13 @@
+package uk.gov.hmcts.reform.professionalapi.domain;
+
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@NoArgsConstructor
+public class Jurisdiction {
+
+    private String id;
+}
