+++ src/main/java/uk/gov/hmcts/reform/professionalapi/service/JurisdictionService.java
@@ -2,7 +2,7 @@ package uk.gov.hmcts.reform.professionalapi.service;
 
 import java.util.List;
 
-import uk.gov.hmcts.reform.professionalapi.controller.request.Jurisdiction;
+import uk.gov.hmcts.reform.professionalapi.domain.Jurisdiction;
 import uk.gov.hmcts.reform.professionalapi.domain.ProfessionalUser;
 
 public interface JurisdictionService {
