+++ src/test/java/uk/gov/hmcts/reform/professionalapi/controller/external/ProfessionalExternalUserControllerTest.java
@@ -25,7 +25,7 @@ import org.springframework.security.core.context.SecurityContextHolder;
 import uk.gov.hmcts.reform.auth.checker.spring.serviceanduser.ServiceAndUserDetails;
 import uk.gov.hmcts.reform.professionalapi.controller.request.InvalidRequest;
 import uk.gov.hmcts.reform.professionalapi.controller.request.OrganisationCreationRequestValidator;
-import uk.gov.hmcts.reform.professionalapi.controller.request.OrganisationIdentifierIdentifierValidatorImpl;
+import uk.gov.hmcts.reform.professionalapi.controller.request.OrganisationIdentifierValidatorImpl;
 import uk.gov.hmcts.reform.professionalapi.controller.request.ProfessionalUserReqValidator;
 import uk.gov.hmcts.reform.professionalapi.domain.Organisation;
 import uk.gov.hmcts.reform.professionalapi.domain.OrganisationStatus;
@@ -41,7 +41,7 @@ public class ProfessionalExternalUserControllerTest {
     private ProfessionalUserService professionalUserServiceMock;
     private Organisation organisation;
     private ProfessionalUserReqValidator profExtUsrReqValidator;
-    private OrganisationIdentifierIdentifierValidatorImpl organisationIdentifierValidatorImpl;
+    private OrganisationIdentifierValidatorImpl organisationIdentifierValidatorImpl;
     private OrganisationCreationRequestValidator organisationCreationRequestValidator;
     private ResponseEntity<?> responseEntity;
 
@@ -54,7 +54,7 @@ public class ProfessionalExternalUserControllerTest {
         organisation = mock(Organisation.class);
         organisationServiceMock = mock(OrganisationService.class);
         professionalUserServiceMock = mock(ProfessionalUserService.class);
-        organisationIdentifierValidatorImpl = mock(OrganisationIdentifierIdentifierValidatorImpl.class);
+        organisationIdentifierValidatorImpl = mock(OrganisationIdentifierValidatorImpl.class);
         profExtUsrReqValidator = mock(ProfessionalUserReqValidator.class);
         organisationCreationRequestValidator = mock(OrganisationCreationRequestValidator.class);
         responseEntity = mock(ResponseEntity.class);
