+++ src/test/java/uk/gov/hmcts/reform/professionalapi/controller/request/JurisdictionUserCreationRequestTest.java
@@ -8,6 +8,7 @@ import java.util.List;
 import java.util.UUID;
 
 import org.junit.Test;
+import uk.gov.hmcts.reform.professionalapi.domain.Jurisdiction;
 
 public class JurisdictionUserCreationRequestTest {
