+++ src/test/java/uk/gov/hmcts/reform/professionalapi/controller/response/MultipleUserProfilesResponseTest.java
@@ -7,6 +7,7 @@ import java.util.ArrayList;
 import java.util.List;
 
 import org.junit.Test;
+import uk.gov.hmcts.reform.professionalapi.domain.UserProfile;
 
 public class MultipleUserProfilesResponseTest {
