+++ src/test/java/uk/gov/hmcts/reform/professionalapi/controller/response/OrganisationsDetailResponseTest.java
@@ -1,0 +1,20 @@
+package uk.gov.hmcts.reform.professionalapi.controller.response;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+import java.util.Collections;
+import org.junit.Test;
+import uk.gov.hmcts.reform.professionalapi.domain.Organisation;
+import uk.gov.hmcts.reform.professionalapi.domain.OrganisationStatus;
+
+public class OrganisationsDetailResponseTest {
+
+    @Test
+    public void test_OrganisationsDetailResponse() {
+        Organisation organisation = new Organisation("Org-Name", OrganisationStatus.PENDING, "sra-id", "companyN", false, "www.org.com");
+
+        OrganisationsDetailResponse organisationsDetailResponseResponse = new OrganisationsDetailResponse(Collections.singletonList(organisation), false);
+
+        assertThat(organisationsDetailResponseResponse).isNotNull();
+    }
+}
