+++ src/test/java/uk/gov/hmcts/reform/professionalapi/domain/RoleAdditionResponseTest.java
@@ -1,0 +1,16 @@
+package uk.gov.hmcts.reform.professionalapi.domain;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+import org.junit.Test;
+
+public class RoleAdditionResponseTest {
+
+    @Test
+    public void addRoleResponseTest() {
+        RoleAdditionResponse addRoleResponse = new RoleAdditionResponse("Code","Message");
+
+        assertThat(addRoleResponse.getIdamStatusCode()).isEqualTo("Code");
+        assertThat(addRoleResponse.getIdamMessage()).isEqualTo("Message");
+    }
+}
