+++ src/test/java/uk/gov/hmcts/reform/professionalapi/helper/OrganisationFixtures.java
@@ -12,8 +12,8 @@ import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
 
-import uk.gov.hmcts.reform.professionalapi.controller.request.Jurisdiction;
 import uk.gov.hmcts.reform.professionalapi.controller.request.OrganisationCreationRequest;
+import uk.gov.hmcts.reform.professionalapi.domain.Jurisdiction;
 
 public class OrganisationFixtures {
